+++ CHANGELOG.md
@@ -2,6 +2,10 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+### Enhancements
+
+* Forbid indirect memory access with negative indices
+
 # v22.9.12
 
 ### Bugfixes